Dynamically switching to new (KLayout 0.28.4) connected-mode DRC checks
This patch is backward-compatible and should provide some performance
improvements plus remove the constraints of the current implementation.
Specifically it will do a true "connected-only" check which means that
space(connected) does not have to be less than space(not connected) and
the DRC descriptions start making sense (e.g. merge nwells).